requirements.txt
discord.py~=2.7.1
PyYaml~=6.0.3
pymongo~=4.16.0
shortuuid~=1.0.13
aiohttp~=3.13.3
jsonschema~=4.26.0
titlecase~=2.4.1
redis~=7.2.1
fluent.runtime~=0.4.0